third_party/WebKit/LayoutTests/fast/dom/Geolocation/not-enough-arguments-expected.txt

Tests how Geolocation handles receiving too few arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.navigator.geolocation.getCurrentPosition() threw exception TypeMismatchError: Failed to execute 'getCurrentPosition' on 'Geolocation': The 1st argument provided is either null, or an invalid Function object..
PASS window.navigator.geolocation.watchPosition() threw exception TypeMismatchError: Failed to execute 'watchCurrentPosition' on 'Geolocation': The 1st argument provided is either null, or an invalid Function object..
PASS window.navigator.geolocation.clearWatch() threw exception TypeError: Failed to execute 'clearWatch' on 'Geolocation': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
